AI_POC/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
numpy
pyarrow
plotly
altair
vega-datasets
streamlit-plotly-events
requests